Handle initial times. Update load_schedules_config
1. Previously schedules are saved into storage, but not queried
anywhere. This commit removes saving. Though persist_schedule() is still
there.

2. Scheduling only starts once the previous schedule times are stored.
Currently this is delayed till _redis connection is up_. This behaviour needs
to be changed if the intended behaviour is to consider jobs from the time
we schedules are loaded (~ server start time) not from the time redis is up.